QILIN-43: Move ui-kit to new repo
